requirements.txt
pypdf==5.1.0
langchain-community==0.3.3
pdf2image==1.17.0
pytesseract==0.3.13
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
ollama==0.3.3
